+++ diesel/src/macros/mod.rs
@@ -458,3 +458,4 @@ macro_rules! print_sql {
 #[macro_use] mod parse;
 #[macro_use] mod query_id;
 #[macro_use] mod insertable;
+#[macro_use] mod queryable;